requirements.txt
speedtest-cli
influxdb-client
pyyaml
requests
psutil